Increase threshold for time since last farming event
The 30 second threshold is violated multiple times per day in the
current network. It's not exactly clear to me why but further analysis
of the logs shows that no signage points are skipped which indicates
that the farmer can still participate in all challenges.

Chiadog is now monitoring signage points so we can safely increase
the threshold here and rely more on other checks.