Notifier: Fix uncaught exceptions and set timeouts
If the internet connection is interrupted and a notifier attempts to
send a notification to the user, it'll fail with uncaught exception
which prevents the watchdog process from further attempts.

This applies a critical fix that will prevent that.

Also add explicit timeouts for all network requests and switch to a less
confusing return value. True = Success instead of True == Failure.