Handlers: Add wallet log handler for added coins
This will replace the notifications for found proof because finding a
proof is not a guarantee for a farmer reward. The found proofs will
still be tracked and reported in the daily summary.

This also has the side-effect that it'll probably also work with
transactions and not only farmer rewards.