NonDecreasingPlots: Increase tolerance threshold to 2
Closes #147
Closes #58

I see little value in making refactors and adding additional
configuration parameters for this. A threshold of 2 should resolve all
issues that are experienced during copying of plots.

Note that 2 is an instantenous threshold since the max value gets reset
after that. Meaning that plot count may drop by 10 without a
notification as long as its dropping in steps of 1.

For transparency of what is happening, I've added an info log that is
always triggered, regardless of the threshold.

The test case is adjusted to contain a scenario with decreasing plot
count by 1 which is expected not to trigger a notification.